// web/default/src/features/wallet/types.ts
export interface UsdtPaymentDetails {
  trade_no: string
  network: string
  address: string
  amount: string
  currency: string
  instructions?: string
}